Region to role descriptions; loc updated in README
One role for each Lambda function is created for each region, and
because we avoid custom physical resource names in CloudFormaton
(hampers resource modification), the region can't be included in the
role name. To help users distinguish these roles, we move the
region from a courtesy tag to the role description. In the list of
roles in the AWS Console, the description column can be added, but
tags can't be.